rename package to mp3convert
