+++ mp3convert/mp3convert/ledger.py
@@ -4,8 +4,8 @@ import logging
 import os
 import shutil
 
-from mp3convertpackage.utils import convert_to_mp3
-from mp3convertpackage.musicfile import MusicFile
+from mp3convert.utils import convert_to_mp3
+from mp3convert.musicfile import MusicFile
 
 logger = logging.getLogger("mp3convert.ledger")
