+++ mp3convert/mp3convert/musicfile.py
