--- mp3convert/mp3convertpackage/__init__.py
@@ -1,5 +1,0 @@
-from mp3convertpackage.ledger import Ledger
-from mp3convertpackage.utils import convert_to_mp3
-from mp3convertpackage.musicfile import MusicFile
-
-__all__ = [Ledger, convert_to_mp3, MusicFile]
