+++ mp3convert/tests/test_mp3convert.py
@@ -1,7 +1,7 @@
 # Package import pattern REF: https://docs.python-guide.org/writing/structure/#test-suite
-from context import mp3convertpackage
+from context import mp3convert
 
-from mp3convertpackage import Ledger
+from mp3convert import Ledger
 
 # content of test_sample.py
 def func(x):
